+++ sdk/expo/android/src/main/java/expo/modules/moengage/internal/Constants.kt
@@ -1,0 +1,3 @@
+package expo.modules.moengage.internal
+
+internal const val BASE_TAG = "MoEngage"