+++ sdk/core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Release Version
 
+- [minor] Added support to include optional module using configuration in package json file
 - Added support for Simplify Inapp Triggers.
 - Android
     - [major] Min SDK version updated to `23`